+++ server-data/resources/[bpt_addons]/bpt_bakerjob/server/main.lua
@@ -23,3 +23,58 @@ ESX.RegisterServerCallback("bpt_bakerjob:SpawnVehicle", function(source, cb, mod
     end)
     cb()
 end)
+
+RegisterNetEvent('bpt_bakerjob:getStockItem')
+AddEventHandler('bpt_bakerjob:getStockItem', function(itemName, count)
+    local xPlayer = ESX.GetPlayerFromId(source)
+
+    if xPlayer.job.name == 'baker' then
+        TriggerEvent('esx_addoninventory:getSharedInventory', 'society_baker', function(inventory)
+            local item = inventory.getItem(itemName)
+
+            -- is there enough in the society?
+            if count > 0 and item.count >= count then
+                -- can the player carry the said amount of x item?
+                if xPlayer.canCarryItem(itemName, count) then
+                    inventory.removeItem(itemName, count)
+                    xPlayer.addInventoryItem(itemName, count)
+                    xPlayer.showNotification(_U('have_withdrawn', count, item.label))
+                else
+                    xPlayer.showNotification(_U('player_cannot_hold'))
+                end
+            else
+                xPlayer.showNotification(_U('quantity_invalid'))
+            end
+        end)
+    else
+        print(('[^3WARNING^7] Player ^5%s^7 attempted ^5bpt_bakerjob:getStockItem^7 (cheating)'):format(source))
+    end
+end)
+
+ESX.RegisterServerCallback('bpt_bakerjob:getStockItems', function(_, cb)
+    TriggerEvent('esx_addoninventory:getSharedInventory', 'society_baker', function(inventory)
+        cb(inventory.items)
+    end)
+end)
+
+RegisterNetEvent('bpt_bakerjob:putStockItems')
+AddEventHandler('bpt_bakerjob:putStockItems', function(itemName, count)
+    local xPlayer = ESX.GetPlayerFromId(source)
+	local sourceItem = xPlayer.getInventoryItem(itemName)
+
+    if xPlayer.job.name == 'baker' then
+        TriggerEvent('esx_addoninventory:getSharedInventory', 'society_baker', function(inventory)
+            local item = inventory.getItem(itemName)
+
+            if sourceItem.count >= count and count > 0 then
+                xPlayer.removeInventoryItem(itemName, count)
+                inventory.addItem(itemName, count)
+                xPlayer.showNotification(_U('have_deposited', count, item.label))
+            else
+                xPlayer.showNotification(_U('quantity_invalid'))
+            end
+        end)
+    else
+        print(('[^3WARNING^7] Player ^5%s^7 attempted ^5bpt_bakerjob:putStockItems^7 (cheating)'):format(source))
+    end
+end)