+++ .github/CHANGELOG.md
@@ -73,3 +73,4 @@ fix lint error: unused argument last; accessing undefined variable Invoke; unuse
 46. [npwd]: fix: Warning: could not find @bitpredator
 47. [server.cfg]: fix: Couldn't find resource @bitpredator
 48. fix: No such export getSharedObject in resource es_extended @bitpredator
+49. [bpt_bakerjob]: refactor: bpt_bakerjob inventory implementation @bitpredator